refactor(core): move app_logger under core.logging and update imports
