+++ src/photo_insight/CSVBatchProcessor.py
@@ -1,7 +1,7 @@
 import os
 import time
 from file_handler.csv_file_handler import CSVFileHandler
-from photo_insight.utils.app_logger import Logger  # ログクラスをインポート
+from photo_insight.core.logging import Logger  # ログクラスをインポート
 from photo_insight.core.batch_framework.base_batch import BaseBatchProcessor
 from composition_evaluator import (
     CompositionEvaluator,
