+++ src/photo_insight/core/logging/__init__.py
@@ -1,0 +1,1 @@
+from .app_logger import *  # noqa: F403
