+++ src/photo_insight/core/logging/app_logger.py
