+++ tests/unit/portrait_quality/test_portrait_quality_evaluator.py
@@ -5,7 +5,7 @@ import pytest
 from photo_insight.evaluators.portrait_quality.portrait_quality_evaluator import (
     PortraitQualityEvaluator,
 )
-from photo_insight.utils.app_logger import Logger
+from photo_insight.core.logging import Logger
 
 ASSETS_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "../../assets"))
